+++ application/models/Partsdata.php
@@ -11,8 +11,10 @@
  *
  * @author Jake
  */
-class Partsdata extends CI_Model {
 
+class Partsdata extends MY_Model
+{
+    
     var $data = array(
         array('id' => '1',
             'name' => 'part A-1',
@@ -195,15 +197,21 @@ class Partsdata extends CI_Model {
             'fileName' => 'w3.jpeg',
             'href' => '/parts/18'),
     );
-
+     
+     
+    
+    
     // Constructor    
-    public function __construct() {
-        parent::__construct();
+    public function __construct()
+    {
+        parent::__construct('partsdata', 'id');
     }
 
     // retrieve all of the parts
-    public function getAllParts() {
-        return $this->data;
+    public function getAllParts()
+    {
+        return $this->all();
+
     }
 
     // retrieve a single part