+++ changelog.md
@@ -60,9 +60,6 @@
 - Redesign and applied bootstrap (Jake)
 - Applied Allman code style (Jake)
 
-### *v.018* (Feb 11, 2017)
-- Added comments to all the history pages (Karan)
-
 ### *v.019* (Feb 11, 2017)
 - Reformatted code (Karan)
 
@@ -74,3 +71,7 @@
 
 ### *v.021* (March 21, 2017)
 - Add user roles (Jake)
+- Add database for parts page (Jake) 
+  - (Config - autoload, constants)
+  - (Models - Partsdata)
+  - (Views - _cell)